+++ src/DotNetStac/Extensions/Disaster/DisastersCharterStacExtension.cs
@@ -8,34 +8,6 @@ using System.Collections.Generic;
 namespace Stac.Extensions.Disaster
 {
     /// <summary>
-    /// Extension methods for accessing EO extension
-    /// </summary>
-    public static class DisasterStacExtensionExtensions
-    {
-        /// <summary>
-        /// Initilize a EoStacExtension class from a STAC item
-        /// </summary>
-        public static DisastersCharterStacExtension DisasterExtension(this IStacObject stacObject)
-        {
-            return new DisastersCharterStacExtension(stacObject);
-        }
-
-        /// <summary>
-        /// Initilize a DisastersCharterStacExtension
-        /// </summary>
-        /// <param name="disasterStacExtension"></param>
-        /// <param name="disastersItemClass"></param>
-        /// <param name="activationId"></param>
-        /// <param name="callIds"></param>
-        public static void Init(this DisastersCharterStacExtension disasterStacExtension, DisastersItemClass disastersItemClass, int activationId, int[] callIds)
-        {
-            disasterStacExtension.Class = disastersItemClass;
-            disasterStacExtension.ActivationId = activationId;
-            disasterStacExtension.CallIds = callIds;
-        }
-    }
-
-    /// <summary>
     /// Disasters Charter Extension to the SpatioTemporal Asset Catalog (STAC) specification.
     /// This extension provides with:
     /// * Additional fields for common disaster properties such as type (e.g. cyclone, earthquake, flooding...).
@@ -44,13 +16,12 @@ namespace Stac.Extensions.Disaster
     /// </summary>
     public class DisastersCharterStacExtension : StacPropertiesContainerExtension, IStacExtension
     {
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
         // Current schema url
         public const string JsonSchemaUrl = "https://terradue.github.io/stac-extensions-disaster/v1.0.0/schema.json";
 
         // Activation id field name
         public const string ActivationIdField = "disaster:activation_id";
-
-        private static IDictionary<string, Type> itemFields;
 
         // Call ids field name
         public const string CallIdsField = "disaster:call_ids";
@@ -72,33 +43,33 @@ namespace Stac.Extensions.Disaster
 
         // Resolution class field name
         public const string ResolutionClassField = "disaster:resolution_class";
+#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
+
+        private readonly IDictionary<string, Type> _itemFields;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DisastersCharterStacExtension"/> class.
         /// Initializes the <see cref="DisastersCharterStacExtension"/> class.
         /// </summary>
-        /// <param name="stacpropertiesContainer"></param>
+        /// <param name="stacpropertiesContainer">The stac properties container.</param>
         public DisastersCharterStacExtension(IStacPropertiesContainer stacpropertiesContainer)
             : base(JsonSchemaUrl, stacpropertiesContainer)
         {
-            if (itemFields == null)
-            {
-                itemFields = new Dictionary<string, Type>();
-                itemFields.Add(ActivationIdField, typeof(int));
-                itemFields.Add(CallIdsField, typeof(int[]));
-                itemFields.Add(CountryField, typeof(string));
-                itemFields.Add(RegionField, typeof(IEnumerable<string>));
-                itemFields.Add(TypeField, typeof(IEnumerable<DisastersType>));
-                itemFields.Add(ClassField, typeof(DisastersItemClass?));
-                itemFields.Add(ActivationStatusField, typeof(DisastersActivationStatus?));
-                itemFields.Add(ResolutionClassField, typeof(DisastersResolutionClass?));
-            }
+            this._itemFields = new Dictionary<string, Type>();
+            this._itemFields.Add(ActivationIdField, typeof(int));
+            this._itemFields.Add(CallIdsField, typeof(int[]));
+            this._itemFields.Add(CountryField, typeof(string));
+            this._itemFields.Add(RegionField, typeof(IEnumerable<string>));
+            this._itemFields.Add(TypeField, typeof(IEnumerable<DisastersType>));
+            this._itemFields.Add(ClassField, typeof(DisastersItemClass?));
+            this._itemFields.Add(ActivationStatusField, typeof(DisastersActivationStatus?));
+            this._itemFields.Add(ResolutionClassField, typeof(DisastersResolutionClass?));
         }
 
         /// <summary>
         /// Gets potential fields and their types
         /// </summary>
-        public override IDictionary<string, Type> ItemFields => itemFields;
+        public override IDictionary<string, Type> ItemFields => this._itemFields;
 
         /// <summary>
         /// Gets or sets identifier of the related Activation
@@ -150,6 +121,7 @@ namespace Stac.Extensions.Disaster
                 {
                     throw new ArgumentException("Country must be a valid ISO-3166 Alpha-3 code");
                 }
+
                 this.StacPropertiesContainer.SetProperty(CountryField, value);
                 this.DeclareStacExtension();
             }
@@ -177,17 +149,23 @@ namespace Stac.Extensions.Disaster
         /// </summary>
         public IEnumerable<DisastersType> Types
         {
-            get { try
-{
-    return (IEnumerable<DisastersType>)this.StacPropertiesContainer.GetProperty<IEnumerable<DisastersType>>(TypeField);
-}
-catch
-{
-    return null;
-}
-}
-            set { this.StacPropertiesContainer.SetProperty(TypeField, value);
-                this.DeclareStacExtension(); }
+            get
+            {
+                try
+                {
+                    return (IEnumerable<DisastersType>)this.StacPropertiesContainer.GetProperty<IEnumerable<DisastersType>>(TypeField);
+                }
+                catch
+                {
+                    return null;
+                }
+            }
+
+            set
+            {
+                this.StacPropertiesContainer.SetProperty(TypeField, value);
+                this.DeclareStacExtension();
+            }
         }
 
         /// <summary>
