+++ src/DotNetStac/Extensions/Disaster/DisastersResolutionClass.cs
@@ -9,6 +9,9 @@ namespace Stac.Extensions.Disaster
     /// </summary>
     public enum DisastersResolutionClass
     {
+#pragma warning disable SA1300 // Element should begin with upper-case letter
+#pragma warning disable SA1602 // EnumerationItemsMustBeDocumented
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
         VLR,
 
         LR,
@@ -18,5 +21,8 @@ namespace Stac.Extensions.Disaster
         HR,
 
         VHR,
+#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
+#pragma warning restore SA1602 // EnumerationItemsMustBeDocumented
+#pragma warning restore SA1300 // Element should begin with upper-case letter
     }
 }
